+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased] - ReleaseDate
 
+## [1.0.9] - 2026-05-06
+
 ### Fixed
 
 - Conditions Observed Generation support.
@@ -374,7 +376,9 @@
 
 <!-- next-url -->
 
-[Unreleased]: https://github.com/agabani/tor-operator/compare/v1.0.8...HEAD
+[Unreleased]: https://github.com/agabani/tor-operator/compare/v1.0.9...HEAD
+
+[1.0.9]: https://github.com/agabani/tor-operator/compare/v1.0.8...v1.0.9
 [1.0.8]: https://github.com/agabani/tor-operator/compare/v1.0.7...v1.0.8
 [1.0.7]: https://github.com/agabani/tor-operator/compare/v1.0.6...v1.0.7
 [1.0.6]: https://github.com/agabani/tor-operator/compare/v1.0.5...v1.0.6